+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 # Fix
 - Removed required fields from client update (#144, PLUM Sprint 230113)
+- Store client cookie domain (#147, PLUM Sprint 230113)
+
+### Features
+- Allow unsetting some client features (#148, PLUM Sprint 230113)
+
 
 ### Refactoring
 - Regex validation of cookie_domain client attribute (#144, PLUM Sprint 230113)